Commence restifying
Todo: don't try creating tables on each request. sqlalchemy doesn't seem
to emit 'create table if exists' sql and it doesn't otherwise seem to lock on the
DB file for sqlite when running create statements from different threads
(which seems to happen when flask handles multiple requests in a row).