Improve binance import - coin/coin pt4 (#70)
- Import fields properly from binance CSV.
- better tests for binance CSV format

Note: tax calculations for transactions imported via the binance parser will not work properly yet. This is because we don't have cost basis amounts (they're not included in the binance report). for example: we just know that the user acquired 1 bitcoin for 125 ethereum, we don't have profit or cost basis in USD.

To turn binance live, we need historical price data to get basis and value for trades.

# Test plan
CI